requirements.txt
requests~=2.26.0
lxml~=4.6.3
pandas~=1.3.3
python-dateutil~=2.8.2
ftfy~=6.0.3
